refactor : main service interface생성
