+++ src/main/java/racinggame/controller/RacingGameController.java
@@ -1,10 +1,11 @@
 package racinggame.controller;
 
-import racinggame.domain.RacingGameService;
+import racinggame.service.RacingGameService;
+import racinggame.service.impl.RacingGameServiceImpl;
 
 public class RacingGameController {
     public void start() {
-        RacingGameService racingGameService = new RacingGameService();
+        RacingGameService racingGameService = new RacingGameServiceImpl();
         racingGameService.startGame();
     }
 }
