+++ src/main/java/racinggame/service/impl/RacingGameServiceImpl.java
@@ -1,0 +1,10 @@
+package racinggame.service.impl;
+
+import racinggame.service.RacingGameService;
+
+public class RacingGameServiceImpl implements RacingGameService {
+
+    public void startGame() {
+
+    }
+}
